Add files via upload
A Project focusing in Auction System Online Interfaces in Admin or Customer User. Focusing the core of the project which is data management and dynamically changes using the data structure of Priority Queues. This commit is still in-progress as forked in a week and now focusing an individual improvement as student that will implement a databases as core of the system flow using MySQL.